requirements.txt
timm>=0.6.13
scipy==1.7.3
yacs==0.1.8
termcolor==2.1.1
einops==0.6.0
numpy>=1.14.6
pandas==1.3.5
tensorboard==2.11.2
matplotlib==3.5.3
seaborn==0.12.2
scikit-learn>=1.0.2
Pillow==9.4.0
torchsummaryX
PyYAML==6.0